Applying patch that reduces compile warnings and fixes warnings from gcc and clang.
Patch contributed by: Gareth (gareth@oneroi.org.uk)

+++ misc.h
@@ -24,8 +24,8 @@ struct buffer
     void *rstart;
     void *rend;
     void *start;
-    int len;
-    int maxlen;
+    size_t len;
+    size_t maxlen;
 #if 0
     unsigned int addr;
     int port;
